refactor:directory changes
